requirements.txt
# --- Framework Web y Servidor ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# --- Bases de Datos (Asíncronas) ---
# PostgreSQL
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# MongoDB (Versiones sincronizadas para evitar ImportError)
motor==3.4.0
pymongo==4.6.3
pydantic==2.5.3
pydantic-settings==2.1.0

# --- Motor de IA y OCR ---
# Pinned a 2.6.1 para mejor soporte de Python 3.12
paddlepaddle-gpu==2.6.1
paddleocr==3.4.0
pymupdf==1.23.8

# --- Procesamiento de Imágenes y Texto ---
opencv-python==4.9.0.80
# CRÍTICO: Numpy debe ser < 2.0.0 para ser compatible con PaddlePaddle
numpy==1.26.4
python-dotenv==1.0.0
# Dependencias necesarias para imgaug/paddle
scipy==1.12.0
imgaug==0.4.0

# --- Generación de E-books ---
EbookLib==0.18
lxml==5.1.0
pypandoc==1.12

# --- Utilidades ---
loguru==0.7.2
psutil==5.9.8